+++ src/main/java/cam72cam/mod/render/cutter/Polygon.java
@@ -2,6 +2,7 @@ package cam72cam.mod.render.cutter;
 
 import cam72cam.mod.math.Vec3d;
 import cam72cam.mod.render.cutter.adapter.QuadTemplate;
+import org.apache.commons.lang3.tuple.Pair;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -9,6 +10,7 @@ import java.util.Collections;
 import java.util.List;
 
 public final class Polygon {
+    private static final double EPS = 1E-6;
     private final List<ClipVertex> vertices;
     private final Vec3d normal; // Could be null
 
@@ -68,6 +70,114 @@ public final class Polygon {
 
         Collections.reverse(vertices);
         return new Polygon(vertices, polygon.getNormal());
+    }
+
+    /**
+     * Splits a convex polygon into quads using fan triangulation from the first vertex.
+     * Triangles are represented as degenerate quads (last vertex duplicated).
+     * Assumes input polygon is convex; for concave polygons, use ear clipping first.
+     *
+     * @param polygon the convex polygon to split
+     * @return a list of quads (each as a Polygon with 4 vertices)
+     */
+    public static List<Polygon> build(Polygon polygon) {
+        List<Polygon> result = new ArrayList<>();
+        List<ClipVertex> verts = polygon.getVertices();
+        int n = verts.size();
+        if (n < 3) return result;
+
+        ClipVertex first = verts.get(0);
+        for (int i = 1; i < n - 1; i++) {
+            ClipVertex a = first;
+            ClipVertex b = verts.get(i);
+            ClipVertex c = verts.get(i + 1);
+            List<ClipVertex> quadVerts = new ArrayList<>(4);
+            quadVerts.add(a);
+            quadVerts.add(b);
+            quadVerts.add(c);
+            quadVerts.add(c);
+            result.add(new Polygon(quadVerts, polygon.getNormal()));
+        }
+        return result;
+    }
+
+    static ClipVertex intersection(
+            ClipVertex a,
+            ClipVertex b,
+            double da,
+            double db) {
+
+        double t = da / (da - db);
+        return a.lerp(b, t);
+    }
+
+    /**
+     * Keep the positive side of the plane.
+     * Returns ClipResult containing clipped polygon and list of (exit, entry) intersection pairs.
+     */
+    public static ClipResult clip(Polygon polygon, Plane plane) {
+        List<ClipVertex> vertices = polygon.getVertices();
+
+        if (vertices.isEmpty()) {
+            return new ClipResult(
+                    new Polygon(new ArrayList<>(), polygon.getNormal()),
+                    new ArrayList<>()
+            );
+        }
+
+        List<ClipVertex> clippedVerts = new ArrayList<>();
+        List<Pair<ClipVertex, ClipVertex>> intersectionPairs = new ArrayList<>();
+        ClipVertex lastExit = null;
+        ClipVertex firstEntry = null; // for wrap-around pairing
+
+        int size = vertices.size();
+
+        for (int i = 0; i < size; i++) {
+            ClipVertex current = vertices.get(i);
+            ClipVertex next = vertices.get((i + 1) % size);
+
+            double dc = plane.distance(current.pos);
+            double dn = plane.distance(next.pos);
+
+            boolean currentInside = dc >= -EPS;
+            boolean nextInside = dn >= -EPS;
+
+            if (currentInside && nextInside) {
+                clippedVerts.add(next);
+
+            } else if (currentInside) {
+                // inside → outside : exit point
+                ClipVertex inter = intersection(current, next, dc, dn);
+                clippedVerts.add(inter);
+                lastExit = inter;
+
+            } else if (nextInside) {
+                // outside → inside : entry point
+                ClipVertex inter = intersection(current, next, dc, dn);
+                clippedVerts.add(inter);
+
+                // Pair with last exit if exists
+                if (lastExit != null) {
+                    intersectionPairs.add(Pair.of(lastExit, inter));
+                    lastExit = null;
+                } else {
+                    // No previous exit → this is the first entry (for wrap-around)
+                    if (firstEntry == null) {
+                        firstEntry = inter;
+                    }
+                }
+                clippedVerts.add(next);
+            }
+            // outside → outside: nothing
+        }
+
+        // Handle wrap‑around: if there's an unpaired exit, pair it with the first entry
+        if (lastExit != null && firstEntry != null) {
+            intersectionPairs.add(Pair.of(lastExit, firstEntry));
+        }
+
+        Polygon clippedPolygon = new Polygon(clippedVerts, polygon.getNormal());
+        return new ClipResult(clippedPolygon, intersectionPairs);
     }
 
     public List<ClipVertex> getVertices() {
